pyproject.toml
[tool.poetry]
name = "bible-verses"
version = "0.1.0"
description = ""
authors = ["David Migoya <dmigoya@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
motor = "^2.3.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
